Moves new production features to v1, updates tests, updates docs.
