+++ language/analysis/README.md
@@ -32,14 +32,14 @@ mvn clean compile assembly:single
 ```
 
 We can then run the assembled JAR file with the `java` command. The variable $COMMAND takes
-three values `entities`, `sentiment`, or `syntax`.
+three values `entities`, `entities-sentiment`, `sentiment`, or `syntax`.
 
 ## Basic usage:
 
 ```
 java -cp target/language-entities-1.0-jar-with-dependencies.jar \
     com.google.cloud.language.samples.Analyze \
-    <entities | sentiment | syntax> \
+    <entities | entities-sentiment | sentiment | syntax> \
     <text | GCS path>
 ```
 
@@ -59,6 +59,13 @@ java -cp target/language-entities-1.0-jar-with-dependencies.jar \
     com.google.cloud.language.samples.Analyze \
     sentiment \
     "The quick brown fox jumped over the lazy dog."
+```
+
+Analyze entity sentiment
+```
+java -cp target/language-entities-1.0-jar-with-dependencies.jar \
+  com.google.cloud.language.samples.Analyze entities-sentiment \
+  "There's nothing better than searching for ice cream on Google."
 ```
 
 Analyze syntax
@@ -90,13 +97,6 @@ java -cp target/language-entities-1.0-jar-with-dependencies.jar \
   com.google.cloud.language.samples.AnalyzeBeta \
   sentiment \
   "Der schnelle braune Fuchs sprang über den faulen Hund."
-```
-
-Analyze entity sentiment Beta
-```
-java -cp target/language-entities-1.0-jar-with-dependencies.jar \
-  com.google.cloud.language.samples.AnalyzeBeta entities-sentiment \
-  "There's nothing better than searching for ice cream on Google."
 ```
 
 Analyze categories in text Beta
