+++ language/analysis/pom.xml
@@ -29,7 +29,6 @@ limitations under the License.
 
   <dependencies>
     <!-- [START dependencies] -->
-    <!-- TODO replace with release version -->
     <dependency>
       <groupId>com.google.cloud</groupId>
       <artifactId>google-cloud-language</artifactId>
