+++ language/analysis/src/main/java/com/google/cloud/language/samples/Analyze.java
@@ -18,6 +18,8 @@ package com.google.cloud.language.samples;
 
 import com.google.cloud.language.v1.AnalyzeEntitiesRequest;
 import com.google.cloud.language.v1.AnalyzeEntitiesResponse;
+import com.google.cloud.language.v1.AnalyzeEntitySentimentRequest;
+import com.google.cloud.language.v1.AnalyzeEntitySentimentResponse;
 import com.google.cloud.language.v1.AnalyzeSentimentResponse;
 import com.google.cloud.language.v1.AnalyzeSyntaxRequest;
 import com.google.cloud.language.v1.AnalyzeSyntaxResponse;
@@ -70,6 +72,12 @@ public class Analyze {
         analyzeSyntaxFile(text);
       } else {
         analyzeSyntaxText(text);
+      }
+    } else if (command.equals("entities-sentiment")) {
+      if (text.startsWith("gs://")) {
+        entitySentimentFile(text);
+      } else {
+        entitySentimentText(text);
       }
     }
   }
@@ -280,5 +288,69 @@ public class Analyze {
       return response.getTokensList();
     }
     // [END analyze_syntax_file]
+  }
+  /**
+   * Detects the entity sentiments in the string {@code text} using the Language Beta API.
+   */
+  public static void entitySentimentText(String text) throws Exception {
+    // [START entity_sentiment_text]
+    // Instantiate a beta client : com.google.cloud.language.v1beta2.LanguageServiceClient
+    try (LanguageServiceClient language = LanguageServiceClient.create()) {
+      Document doc = Document.newBuilder()
+          .setContent(text).setType(Type.PLAIN_TEXT).build();
+      AnalyzeEntitySentimentRequest request = AnalyzeEntitySentimentRequest.newBuilder()
+          .setDocument(doc)
+          .setEncodingType(EncodingType.UTF16).build();
+      // detect entity sentiments in the given string
+      AnalyzeEntitySentimentResponse response = language.analyzeEntitySentiment(request);
+      // Print the response
+      for (Entity entity : response.getEntitiesList()) {
+        System.out.printf("Entity: %s\n", entity.getName());
+        System.out.printf("Salience: %.3f\n", entity.getSalience());
+        System.out.printf("Sentiment : %s\n", entity.getSentiment());
+        for (EntityMention mention : entity.getMentionsList()) {
+          System.out.printf("Begin offset: %d\n", mention.getText().getBeginOffset());
+          System.out.printf("Content: %s\n", mention.getText().getContent());
+          System.out.printf("Magnitude: %.3f\n", mention.getSentiment().getMagnitude());
+          System.out.printf("Sentiment score : %.3f\n", mention.getSentiment().getScore());
+          System.out.printf("Type: %s\n\n", mention.getType());
+        }
+      }
+    }
+    // [END entity_sentiment_text]
+  }
+
+  /**
+   * Identifies the entity sentiments in the the GCS hosted file using the Language Beta API.
+   */
+  public static void entitySentimentFile(String gcsUri) throws Exception {
+    // [START entity_sentiment_file]
+    // Instantiate a beta client : com.google.cloud.language.v1beta2.LanguageServiceClient
+    try (LanguageServiceClient language = LanguageServiceClient.create()) {
+      Document doc = Document.newBuilder()
+          .setGcsContentUri(gcsUri)
+          .setType(Type.PLAIN_TEXT)
+          .build();
+      AnalyzeEntitySentimentRequest request = AnalyzeEntitySentimentRequest.newBuilder()
+          .setDocument(doc)
+          .setEncodingType(EncodingType.UTF16)
+          .build();
+      // Detect entity sentiments in the given file
+      AnalyzeEntitySentimentResponse response = language.analyzeEntitySentiment(request);
+      // Print the response
+      for (Entity entity : response.getEntitiesList()) {
+        System.out.printf("Entity: %s\n", entity.getName());
+        System.out.printf("Salience: %.3f\n", entity.getSalience());
+        System.out.printf("Sentiment : %s\n", entity.getSentiment());
+        for (EntityMention mention : entity.getMentionsList()) {
+          System.out.printf("Begin offset: %d\n", mention.getText().getBeginOffset());
+          System.out.printf("Content: %s\n", mention.getText().getContent());
+          System.out.printf("Magnitude: %.3f\n", mention.getSentiment().getMagnitude());
+          System.out.printf("Sentiment score : %.3f\n", mention.getSentiment().getScore());
+          System.out.printf("Type: %s\n\n", mention.getType());
+        }
+      }
+    }
+    // [END entity_sentiment_file]
   }
 }
